test(cli): split tsconfig tests
